feat: print with slashes + add comments
The dir_separator "»" is replaced by a slash when printing or suggesting completion.
Stars should be referenced using slashes, that will automaticaly be replaced by this separator for the real star names. See line 183 for the corresponding comment.
Replacing characters by slashes might not be straightforward sometimes, as some need to be escaped and some don't.
Tested on bash5.2 and zsh5.8.1.